Upload files to configure subject
Checks that the subject is configured
If it is not configured and you are a student, you are asked to wait for instructions from the teacher. If you are a teacher, EDUTrack sends the format of student files and activities for configuration.
If students and/or activities have already been uploaded, it is indicated that the file must be renamed with the prefix "add" to add elements or "replace" to replace all elements. The latter will also delete the grades section, and generate it again with the new data.